v1/requirements.txt
anthropic>=0.39.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
pyautogui>=0.9.54
pillow>=10.1.0
python-multipart>=0.0.6
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.0.0

